requirements.txt
# ====================================
# MindfulLog-AI-Wellness-Journal-RAG
# Dependencies
# ====================================

# --- Web Framework & Server ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# --- Frontend Demo ---
streamlit==1.41.1
requests==2.32.3

# --- AI / LLM ---
openai==1.58.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langchain-text-splitters==0.3.4

# --- Vector Store ---
chromadb==0.5.23

# --- Environment & Utilities ---
python-dotenv==1.0.1
